pyproject.toml
[project]
name = "mcp-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "mcp[cli]>=1.18.0",
    "python-dotenv>=1.1.1",
    "spotipy>=2.25.1",
]
